move distance logic to reducer
