+++ src/components/App/App.jsx
@@ -29,11 +29,11 @@ function App() {
 
   const user = useSelector(store => store.user);
 
-  // useEffect(() => {
-  //   dispatch({ type: 'FETCH_USER' });
-  //   dispatch({ type: 'GET_LIST'});
-  //   console.log('use effect in App ran');
-  // }, []);
+  useEffect(() => {
+    dispatch({ type: 'FETCH_USER' });
+    // dispatch({ type: 'GET_LIST'});
+    console.log('use effect in App ran');
+  }, []);
 
   return (
     <Router>
